+++ packages/app/hivecommand-frontend/src/pages/device-control/utils/program.ts
@@ -141,6 +141,8 @@ import { gql, useApolloClient, useQuery } from "@apollo/client";
                                             script
                                         }
                                     }
+
+                                    options
         
                                     configuration {
                                         id