+++ packages/clients/native-app-client/src/App.tsx
@@ -45,7 +45,6 @@ function App() {
     return () => {
       unlisten();
     }
-
   }, [])
 
   useEffect(() => {