+++ packages/core-ui/command-surface/src/index.tsx
@@ -100,6 +100,9 @@ export interface HMINode {
             }
             value: any
         }[]
+        options?: {
+            [id: string]: any
+        }
     };
 
     // template?: HMITemplate;