+++ packages/editors/interface/src/sidebar/panes/configuration/index.tsx
@@ -7,8 +7,9 @@ import { DataTypes, formatInterface, fromOPCType, lookupType, toJSType } from '@
 import { Node } from 'reactflow';
 import { HMITag, HMITemplate, HMIType } from '@hive-command/interface-types'
 import { useInterfaceEditor } from '../../../context';
+import { ConfigInput } from './render';
+import { TemplateConfiguration } from './templates';
 
-export type ConfigInputType = 'Function' | 'Tag' | 'String' | '[String]' | 'Number' | 'Boolean'
 
 
 export interface ConfigurationPaneProps {
@@ -27,11 +28,9 @@ export const ConfigurationPane: React.FC<ConfigurationPaneProps> = (props) => {
 
     const configuredOptions = selectedNode?.data?.configuredOptions || {};
 
-    const templateOptions = selectedNode?.data?.templateOptions || [];
 
     const options = selectedNode?.data?.options || {};
 
-    const [templateState, setTemplateState] = useState<{ field: { id: string }, value: any }[]>([])
 
     const [state, setState] = useState<{ key: string, value: any }[]>([])
 
@@ -41,7 +40,6 @@ export const ConfigurationPane: React.FC<ConfigurationPaneProps> = (props) => {
     const [functionOpt, setFunctionOpt] = useState<any>()
 
 
-    const [ws, setWs] = useState<{ [key: string]: any }>({});
 
     const updateHMINode = () => {
 
@@ -95,20 +93,6 @@ export const ConfigurationPane: React.FC<ConfigurationPaneProps> = (props) => {
         }
     }, [ JSON.stringify(options), JSON.stringify(configuredOptions) ])
 
-    console.log({
-        state,
-        configuredOptions,
-        options,
-        templateOptions
-    })
-
-    useEffect(() => {
-        if (templateOptions) {
-            setTemplateState(templateOptions);
-        }
-    }, [templateOptions])
-
-
 
     const [updateBouncer, setUpdateBouncer] = useState<any>(null);
 
@@ -133,252 +117,7 @@ export const ConfigurationPane: React.FC<ConfigurationPaneProps> = (props) => {
 
     }
 
-
-    const renderConfigInput = ({ type, value, label, id }: { type: ConfigInputType, value: any, label: string, id?: string }, updateState: ((key: string, value: any) => void) = _updateState) => {
-
-        // label = id || label;
-
-        if (typeof (value) === 'string' && value?.indexOf('script://') > -1 && type !== "Function") {
-            return (<Box sx={{ display: 'flex', flexDirection: 'column' }}>
-                <Typography fontSize="small">{label}</Typography>
-                <Typography color="gray" fontSize={"small"}>Provided by script</Typography>
-
-            </Box>)
-        }
-
-        const type_parts = type.split(':');
-        if (type_parts.length > 1) {
-            type = type_parts[0] as ConfigInputType;
-        }
-
-        switch (type) {
-            case 'Tag':
-
-                let options = tags?.slice()?.sort((a, b) => `${a.name}`.localeCompare(`${b.name}`));
-
-                if (type_parts[1]) {
-                    options = options?.filter((a) => a?.type == types.find((a) => a.id === type_parts[1])?.name);
-                }
-
-
-                return (
-                    // <Box>
-
-                    <Autocomplete
-                        disablePortal
-                        options={options}
-                        value={options?.find((a) => a.id == value) || null}
-
-                        onChange={(event, newValue) => {
-                            if (typeof (newValue) === 'string') return console.error("Tag input with string value")
-                            updateState(id || label, newValue?.id)
-                            // if(!newValue){
-                            //     updateState(label, null)
-                            // }else{
-                            //     setFunctionArgs(newValue);
-                            //     setFunctionOpt(label)
-                            // }
-                        }}
-                        getOptionLabel={(option) => typeof (option) == "string" ? option : `${option.name}`}
-                        // isOptionEqualToValue={(option, value) => option.id == value.id}
-                        renderInput={(params) =>
-                            <TextField
-                                {...params}
-                                label={label}
-                            />
-                        }
-                        // value={value || ''}
-                        // onChange={(e) => {
-                        //     updateState(label, e.target.value)
-                        // }}
-                        size="small"
-                    // label={label} />
-                    />
-
-                    // </Box>
-                )
-            case 'Boolean':
-                return (
-                    <Box sx={{ display: 'flex', alignItems: 'center' }}>
-                        <input
-                            checked={Boolean(value)}
-                            type="checkbox"
-                            onChange={(evt) => {
-                                updateState(id || label, evt.target.checked)
-                            }} />
-
-                        <Typography>{label}</Typography>
-                    </Box>
-                )
-            case 'Function':
-                return (
-                    <Box sx={{ display: 'flex', alignItems: 'center' }}>
-                        <Typography sx={{ flex: 1 }} fontSize={'small'}>{label}</Typography>
-                        <IconButton
-                            size="small"
-                            onClick={() => {
-                                setFunctionOpt(label);
-
-
-                                // const valueState = assignableDevices?.map((dev) => {
-                                //     let state = dev.type.state.map((stateItem) => { 
-                                //         return {key: stateItem.key, value: fromOPCType(stateItem.type)}
-                                //      }).reduce((prev, curr) => ({
-                                //         ...prev,
-                                //         [curr.key]: curr.value
-                                //      }), {})
-
-                                //      return {key: dev.tag, value: state}
-                                // // `${dev.tag}: { ${dev.type?.state?.map((stateItem) => `${stateItem.key}: ${ getOPCType(stateItem.type) }`).join('\n')} }`).join(';\n')
-                                // }).reduce((prev, curr) => ({
-                                //     ...prev,
-                                //     [curr.key]: curr.value
-                                // }), {});
-
-                                console.log("SCRIPT", value);
-
-                                setFunctionArgs({
-                                    defaultValue: value ? value?.match(/script:\/\/(.*)/s)?.[1] : `export const handler = (elem: {x: number, y: number, width: number, height: number}, state: Tags, setState: (values: DeepPartial<Tags>) => void, args: any[], transformer: (state: any) => any) => {
-
-
-}`,
-                                    extraLib: `
-                                declare function showWindow(
-                                    position: {x: number, y: number, width: number, height: number, anchor?: string},
-                                    data: (state: any) => any,
-                                    transformer?: (state: any) => any
-                                ){
-
-                                }
-
-                                declare function showTagWindow(
-                                    position: {x: number, y: number, width: number, height: number, anchor?: string},
-                                    deviceTag: string,
-                                    state: string[],
-                                    actions?: { label: string, func: string }[],
-                                    setpoints?: (state: any) => ({ label: string, getter: () => any, setter?: (value: any) => void }[]),
-                                    manual?: (state: any) => ({getter: () => boolean, setter: (value: boolean) => void}),
-                                    transformer?: (state: any) => any
-                                ){
-                                    
-                                }
-
-                                declare function changeView(view: string){
-
-                                }
-
-                                ${typeSchema}
-
-                                ${formatInterface('Tags', tagSchema)}
-                            `
-                                });
-                            }}>
-                            <Javascript fontSize='inherit' />
-                        </IconButton>
-                    </Box>
-                );
-            case '[String]':
-                return (() => {
-
-
-                    return (<Autocomplete
-                        options={value || []}
-                        size="small"
-                        value={ws?.[label] || ''}
-
-                        renderOption={(props, option) => (
-                            <Box sx={{ display: 'flex', position: 'relative', alignItems: 'center', '& .MuiIconButton-root': { opacity: 0 }, '&:hover .MuiIconButton-root': { opacity: 1 } }}>
-                                <li style={{ flex: 1 }} {...props as any}>
-                                    <Typography>{option}</Typography>
-                                </li>
-                                <Box sx={{
-                                    position: 'absolute',
-                                    right: '6px',
-                                    top: 0,
-                                    bottom: 0,
-                                    display: 'flex',
-                                    alignItems: 'center'
-                                }}>
-                                    <IconButton size="small"><Delete color='error' fontSize='inherit' /></IconButton>
-                                </Box>
-                            </Box>)}
-                        onKeyDown={(e) => {
-                            if (e.key === "Enter") {
-                                updateState(id || label, [...new Set((value || []).concat(ws?.[label]))])
-                                setWs((ws) => { ws[label] = ''; return ws })
-                            }
-                        }}
-                        renderInput={(params) => <TextField {...params} onChange={(e) => {
-                            setWs((ws) => { ws[label] = e.target.value; return ws })
-                        }} label={label} />}
-                        freeSolo
-                    />)
-                })();
-
-            case 'String':
-
-                return (
-                    <TemplateInput
-                        label={label}
-                        value={value || ''}
-                        options={tagInputs}
-                        onChange={(e) => {
-                            updateState(id || label, e)
-                        }}
-                    />
-                );
-            case 'Number':
-                return (
-                    <TextField
-                        value={value || ''}
-                        type="number"
-                        onChange={(e) => {
-                            updateState(id || label, e.target.value)
-                        }}
-                        size="small"
-                        label={label} />
-                );
-
-            default:
-                return (<span>{type} not found in renderConfigInput</span>)
-        }
-    }
-
     const activeTemplate = useMemo(() => templates?.find((a) => a.id === selectedNode?.data?.template), [selectedNode?.data?.template]);
-
-    const templateInputs = useMemo(() => {
-        return activeTemplate ? (
-            <Card elevation={5} sx={{ padding: '6px' }}>
-                <Typography sx={{ marginBottom: '6px' }} fontSize={'small'}>Template Options</Typography>
-                {activeTemplate?.inputs?.map((input) => (
-                    <Box sx={{ marginBottom: '6px' }}>
-                        {renderConfigInput(
-                            { id: input.id, type: input.type, value: templateState?.find((a) => input.id === a.field?.id)?.value || null, label: input.name },
-                            (key, value) => {
-                                props.onNodeUpdate?.({
-                                    data: {
-                                        templateOptions: { [key]: value }
-                                    }
-                                });
-
-                                // props.onNodeTemplateUpdate?.({[key]:  value})
-
-                                // updateNodeTemplateConfig({
-                                //     variables: {
-                                //         nodeId: selected?.id,
-                                //         fieldId: key,
-                                //         value
-                                //     }
-                                // }).then(() => {
-                                //     refetch();
-                                // })
-                            }
-                        )}
-                    </Box>
-                ))}
-            </Card>
-        ) : null
-    }, [activeTemplate, templateState])
 
 
     return (
@@ -417,34 +156,8 @@ export const ConfigurationPane: React.FC<ConfigurationPaneProps> = (props) => {
             // ]}
             />
 
-            <Box sx={{ marginBottom: '6px' }}>
-                <Autocomplete
-                    fullWidth
-                    options={templates || []}
-                    value={templates?.find((a) => a.id === selectedNode?.data?.template) || null}
-                    disablePortal
-                    onChange={(e, newVal) => {
-                        if (typeof (newVal) === 'string') return;
-                        // if(!selected?.id) return
-
-                        props.onNodeUpdate?.({ data: { template: newVal?.id || null } });
-
-                        // assignNodeTemplate({
-                        //     variables: {
-                        //         nodeId: selected?.id,
-                        //         input: {
-                        //             template: newVal?.id || null
-                        //         }
-                        //     }
-                        // }).then(() => refetch?.());
-                    }}
-                    getOptionLabel={(option) => typeof (option) === 'string' ? option : option?.name}
-                    renderInput={(params) => <TextField  {...params} size="small" label="Template" />}
-                />
-
-            </Box>
-            {templateInputs}
-
+         
+            <TemplateConfiguration onNodeUpdate={props.onNodeUpdate} />
 
             <Divider sx={{ marginTop: '12px', marginBottom: '12px' }} />
             <Typography fontSize="small">Node Options</Typography>
@@ -465,7 +178,15 @@ export const ConfigurationPane: React.FC<ConfigurationPaneProps> = (props) => {
                         ) : (
                             <>
                                 <Box sx={{ flex: 1, opacity: templateOutput ? 0.1 : 1 }}>
-                                    {renderConfigInput({ type, value, label })}
+                                    <ConfigInput
+                                        type={ type }
+                                        value={value}
+                                        label={label}
+                                        onUpdateState={_updateState}
+                                        onUpdateStateJs={(key, args) => {
+                                            setFunctionArgs(args || null);
+                                            setFunctionOpt(key)
+                                        }} />
                                 </Box>
                                 {type !== "Function" && <IconButton
                                     onClick={() => {