+++ packages/types/interface-types/src/DataTransformer/index.ts
@@ -1,0 +1,18 @@
+export interface TemplateIO {
+    id: string;
+    name: string;
+    type: string;
+}
+
+export interface TransformerTemplate {
+    inputs?: TemplateIO[]
+    outputs?: TemplateIO[]
+
+    edges?: {to: {id: string}, script: string}[]
+}
+export interface DataTransformer {
+    template: TransformerTemplate
+
+    configuration?: any[];
+    options?: any;
+}