+++ packages/clients/cli-client/CHANGELOG.md
@@ -3,6 +3,46 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.4.15-alpha.285](https://github.com/TheTechCompany/HiveCommand/compare/v1.4.15-alpha.284...v1.4.15-alpha.285) (2023-11-14)
+
+**Note:** Version bump only for package @hive-command/cli-client
+
+
+
+
+
+## [1.4.15-alpha.284](https://github.com/TheTechCompany/HiveCommand/compare/v1.4.15-alpha.283...v1.4.15-alpha.284) (2023-11-13)
+
+**Note:** Version bump only for package @hive-command/cli-client
+
+
+
+
+
+## [1.4.15-alpha.283](https://github.com/TheTechCompany/HiveCommand/compare/v1.4.15-alpha.282...v1.4.15-alpha.283) (2023-11-10)
+
+**Note:** Version bump only for package @hive-command/cli-client
+
+
+
+
+
+## [1.4.15-alpha.282](https://github.com/TheTechCompany/HiveCommand/compare/v1.4.15-alpha.281...v1.4.15-alpha.282) (2023-11-10)
+
+**Note:** Version bump only for package @hive-command/cli-client
+
+
+
+
+
+## [1.4.15-alpha.281](https://github.com/TheTechCompany/HiveCommand/compare/v1.4.15-alpha.280...v1.4.15-alpha.281) (2023-11-10)
+
+**Note:** Version bump only for package @hive-command/cli-client
+
+
+
+
+
 ## [1.4.15-alpha.280](https://github.com/TheTechCompany/HiveCommand/compare/v1.4.15-alpha.279...v1.4.15-alpha.280) (2023-11-10)
 
 **Note:** Version bump only for package @hive-command/cli-client